% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{airpassenger_JFK_MIA}
\alias{airpassenger_JFK_MIA}
\title{Total number of passengers on domestic flights between JFK (New York) and MIA (Miami) airports}
\format{
A data frame with 405 rows and 2 variables:
\describe{
\item{date}{date of the day in YYYY-MM-DD format}
\item{passengers}{total number of passengers}
...
}
}
\source{
US Bureau of Transportation Statistics website:
}
\usage{
Time series analysis. Seasonality. Extreme events (covid-19 pandemic).
}
\description{
The data are monthly observations from January, 1990 to September, 2023.
}
\keyword{datasets}
