% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{campy}
\alias{campy}
\title{Campylobacter infections in northern Quebec}
\format{
A data frame with 140 rows and 1 variable:
\describe{
\item{infections}{Number of campylobacter infections during a four week period.}
...
}
}
\usage{
campy
}
\description{
A dataset of the number of campylobacter infections in the north of the
province Quebec (Canada) in four week intervals from January 1990 to the
end of October 2000. 13 observations per year.
}
\keyword{datasets}
