% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chinagdp}
\alias{chinagdp}
\title{Yearly gross domestic product (GDP) per capita for China 1960-2020 in current US dollar.}
\format{
A data frame with 60 rows and 3 variables:
\describe{
\item{year}{year in YYYY format}
\item{gdp}{gross domestic product per capita in current US dollar}
\item{gdp_growth}{percentage yearly growth}
...
}
}
\source{
The World Bank \url{https://data.worldbank.org/indicator/NY.GDP.PCAP.CD?locations=CN}
}
\usage{
chinagdp
}
\description{
The dataset also contains the yearly growth rate with a missing value for the year 1960.
}
\keyword{datasets}
