% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{debtratio}
\alias{debtratio}
\title{Financial leverage of US firms}
\format{
A data frame with 4405 rows and 5 variables:
\describe{
\item{tang}{tangible assets/book value of total assets}
\item{mbtr}{(book value of total assets − book value of equity + market value of equity)/book value of total assets}
\item{logsale}{natural log of sales}
\item{profit}{earnings before interest, taxes, depreciation, and amortization (EBITDA)/book value of total assets}
\item{debtratio}{total debt/(total debt + book value of equity)}
...
}
}
\source{
Villani, M. Kohn, R and Nott, D.J. (2012). Generalized smooth finite mixtures. Journal of Econometrics, vol. 171 pp. 121–133.
}
\usage{
debtratio
}
\description{
A dataset on the leverage (debtratio) for 4405 US firms.
}
\keyword{datasets}
