% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{energySweden}
\alias{energySweden}
\title{The total energy usage in mega watt hours in Sweden 2021 in units of 10 millions.}
\format{
A data frame with 234 rows and 2 variables:
\describe{
\item{region}{municipalities}
\item{Mwh}{Mega watt hours in Sweden 2021 in units of 10 millions}
...
}
}
\source{
Statistics Sweden, 2023
}
\usage{
energySweden
}
\description{
This data set some missing data and three outliers (of three major municipalities)
}
\keyword{datasets}
