% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fish}
\alias{fish}
\title{Data on fish length}
\format{
A data frame with 44 rows and 4 variables:
\describe{
\item{length}{fish length, in mm}
\item{intercept}{just a vector of ones for the intercept}
\item{age}{fish age, in days}
\item{temp}{temperature of the water, in celcius degrees}
}
}
\source{
A subset of the data in D. M. Titterington, A. F. M. Smith and U.E. Makov (1985) Statistical Analysis of Finite Mixture Distributions. Wiley.
}
\usage{
fish
}
\description{
This dataset contains the length of 44 fishes with varying age and water temperature.
}
\keyword{datasets}
