% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{galton}
\alias{galton}
\title{Data on the height of children and their parents}
\format{
A data frame with 898 rows and 6 variables:
\describe{
\item{Family}{Family identifier}
\item{Father}{Height of father, in inches}
\item{Mother}{Height of mother, in inches}
\item{Gender}{Gender of the child}
\item{Height}{Height of the adult child}
\item{Kids}{Number of children in the family}
}
}
\source{
Marshall Bradley, "Galton Parent and Child Height Data" from the Wolfram Data Repository (2022).
}
\usage{
galton
}
\description{
This dataset contains the height of 898 adult children and their
parents. The data was (perhaps) used in "Regression Towards
Mediocrity in Hereditary Stature" by Francis Galton (1886).
}
\keyword{datasets}
