% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{globaltemp}
\alias{globaltemp}
\title{Global temperature anomaly time series data during 1880-2020.}
\format{
A data frame with 141 rows and 3 variables:
\describe{
\item{year}{year}
\item{temp}{global temperature anomaly (i.e. relative to 1951-1980 average temperatures)}
\item{smoothtemp}{lowess smoothed version of temp}
...
}
}
\source{
NASA \url{https://climate.nasa.gov/vital-signs/global-temperature/}.
}
\usage{
globaltemp
}
\description{
The data contains yearly observations of global surface temperature relative to 1951-1980 average temperatures.
}
\keyword{datasets}
