% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{healthbudget}
\alias{healthbudget}
\title{Health budget and life expectancy in 30 countries.}
\format{
A data frame with 30 rows and 3 variables:
\describe{
\item{country}{Country name}
\item{spending}{Spending on health in thousands of dollars per capita.}
\item{lifespan}{Life expectancy in years}
...
}
}
\source{
Gelman, Hill and Vehtari (2020). Regression and other stories, \emph{Cambridge University Press}. \url{https://avehtari.github.io/ROS-Examples/}
}
\usage{
healthbudget
}
\description{
Health budget and life expectancy in 30 countries.
}
\keyword{datasets}
