% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{maklardata_rot}
\alias{maklardata_rot}
\title{Real estate data from chapter 1 of "Regression och tidsserieanalys" (ÅJA)}
\format{
A data frame with 28 rows and 5 variables:
\describe{
\item{pris}{Cost of the property}
\item{tomt}{Size of the plot}
\item{yta}{Size (in square meters) of the property}
\item{alder}{Age of the property}
\item{antal_rum}{Number of rooms in the property}
}
}
\source{
Andersson, G., Jorner, U., and Ågren, A. "Regression och tidsserieanalys (2007)".
}
\usage{
maklardata_rot
}
\description{
Contains data on the price (pris), size of the plot (tomt), size of
the property (yta), age (alder) and number of rooms (antal_rum) of 28
properties from the book "Regression och Tidsserieanalys".
}
\keyword{datasets}
