% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{orings}
\alias{orings}
\title{Challenger USA Space Shuttle O-Ring Data Set}
\format{
A data frame with 23 rows and 5 variables:
\describe{
\item{nRisk}{Number of O-rings at risk on a given flight}
\item{nDistress}{Number experiencing thermal distress}
\item{tempLaunch}{Launch temperature (degrees F)}
\item{pressure}{Leak-check pressure (psi)}
\item{temporalOrder}{Temporal order of flight}
...
}
}
\source{
Draper, D. (1995). Assessment and propagation of model uncertainty. JRSS B.
\url{https://www.jstor.org/stable/2346087} and the UCI repository
\url{https://archive.ics.uci.edu/ml/datasets/Challenger+USA+Space+Shuttle+O-Ring}
}
\usage{
orings
}
\description{
The number of O-rings that experience thermal distress on 23 shuttle flights.
}
\keyword{datasets}
