% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{salaries}
\alias{salaries}
\title{Salaries of US university professors}
\format{
A data frame with 397 rows (professors) and 13 variables:
\describe{
\item{Rank}{categorical variable for persons academic rank ("AsstProf", "AssocProf", "Prof")}
\item{Discipline}{Scientific discipline "A" or "B"}
\item{YrsSincePhD}{Number of years since PhD degree}
\item{YrsService}{Number of years in service at university}
\item{Sex}{Female/Male}
\item{Salary}{Yearly salary in US dollars}
\item{LogSalary}{Log of Salary}
\item{PhdAge}{Normalized version of YrsSincePhD from 0 to 1 (person with most YrsSincePhD)}
\item{PhdAgeSqr}{Square of PhdAge}
\item{Rank2}{Binary dummy variable for Rank = "AssocProf"}
\item{Rank3}{Binary dummy variable for Rank = "Prof"}
...
}
}
\source{
This is a dataset from R, modified by transformations and creation of dummy variables etc.
}
\usage{
salaries
}
\description{
Salaries of US university professors
}
\keyword{datasets}
