% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{smoking}
\alias{smoking}
\title{Smoking habits}
\format{
A data frame with 60 rows and 4 variables:
\describe{
\item{Gender}{0 = Man, 1 = Woman}
\item{Smoking habits}{0 = no smoking, 1 = yes smoking}
\item{Age}{age of the person}
\item{Training}{number of times a person is training per week}
...
}
}
\source{
Nyquist, H. (2021), Statistikens grunder. Studentlitteratur.
}
\usage{
smoking
}
\description{
This dataset is included in the extra material of the book: "Statistikens grunder". It contains four variables. Each row has the data for a person
by including gender, smoking habits, age and training.
NOTE: this is the same dataset reported in the book (as rokdata.xls), but the name of the variables and the related descriptions are in English.
}
\keyword{datasets}
