% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tableware}
\alias{tableware}
\title{Data on the polishing times for tableware}
\format{
A data frame with 59 rows and 4 variables:
\describe{
\item{price}{Price of the item.}
\item{time}{Total polishing time, measured in minutes.}
\item{type}{Categorical variable that can take on the values "Bowl", "Casserole", "Dish", "Plate", and "Tray".}
\item{diameter}{Diamter of the piece of tableware.}
}
}
\source{
\url{https://dasl.datadescription.com/datafiles/?_sf_s=nambe\%20mills&_sfm_cases=4+59943}
}
\usage{
tableware
}
\description{
One step in the process of manufacturing tableware is polishing.
This dataset contains polishing times, measured in minutes, for a
number of items from a factory in Santa Fe.
}
\keyword{datasets}
