% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tempLinkoping}
\alias{tempLinkoping}
\title{Temperature at Malmslätt, Linköping}
\format{
A data frame with 366 rows and 2 variables:
\describe{
\item{time}{The number of days since the beginning of the year divided
by 366}
\item{temp}{Temperature in Celcius}
...
}
}
\source{
Swedish Meteorological and Hydrological Institute \url{https://www.smhi.se/data/meteorologi/temperatur}
}
\usage{
tempLinkoping
}
\description{
A dataset of daily temperatures (in Celcius) at Malmslätt, Linköping, over
the course of the year 2016 (366 observations since 2016 was a leap year).
}
\keyword{datasets}
