% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{womenwork}
\alias{womenwork}
\title{Data on womens' labor force participation}
\format{
A data frame with 200 rows and 9 variables:
\describe{
\item{work}{Whether or not the woman works.}
\item{constant}{Constant/intercept.}
\item{husbandInc}{Husband's income.}
\item{educYears}{Years of education.}
\item{expYears}{Years of experience.}
\item{expYears2}{(Years of experience/10)^2.}
\item{age}{Age.}
\item{nSmallChild}{Number of children at or under the age of six
in the household.}
\item{nBigChild}{Number of children older than six in the household.
Not counting the husband.}
...
}
}
\source{
Mroz, T. A. (1987). The sensitivity of an empirical model of married women's
hours of work to economic and statistical assumptions. Econometrica 55, 765–799.

This particular data is the subset used in

Jelizakov, I. and Chib, S. (2001). Marginal Likelihood from the Metropolis–Hastings Output. JASA.
}
\usage{
womenwork
}
\description{
A dataset on the working status (works or does not work) for 200 women.
}
\keyword{datasets}
