% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bloodpressure}
\alias{bloodpressure}
\title{Systolic blood pressure for male and females}
\format{
A data frame with 69 rows and 3 variables:
\describe{
\item{SBP}{Systolic blood pressure}
\item{Age}{Age of person}
\item{Gender}{1 = male, 0 = woman}
...
}
}
\source{
???
}
\usage{
bloodpressure
}
\description{
Systolic blood pressure for male and females
}
\keyword{datasets}
